+++ software/alarm_clock_jukebox/Src/main.c
@@ -3,14 +3,14 @@
 int main(void) {
 
   typedef struct {
-    int on;
-    int hour;
-    int minute;
+    uint8_t on;
+    uint8_t hour;
+    uint8_t minute;
   } config_struct;
 
   typedef struct {
-    int display;
-    int alarm;
+    uint8_t display;
+    uint8_t alarm;
     config_struct config;
   } mode_struct;
 
@@ -30,11 +30,6 @@ int main(void) {
     printf("Interrupts successively initialized!\n");
     enableTimerSecondISR();
   }
-
-
-
-
-
 
   while(1) {}  
 }
